norme-errors

+++ ft_strlen.c
@@ -1,11 +1,11 @@
 #include "libft.h"
 
-int ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-    const char *str1;
+	const char *str1;
 
-    str1 = str;
-    while (*str)
-        str++;
-    return (str - str1);
+	str1 = str;
+	while (*str)
+		str++;
+	return (str - str1);
 }
